pyproject.toml
[tool.poetry]
name = "fastapi-cas-example"
version = "0.1.0"
description = "An example repo to demo how to integrate python-cas into a fastapi app."
authors = ["Chise1 <chise123@love.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
fastapi = "^0.63.0"
python-cas = "^1.5.0"
uvicorn = "^0.13.3"
itsdangerous = "^1.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
url = "https://mirrors.aliyun.com/pypi/simple/"
name = "aliyun"
default = true